feat: Добавлена проверка прав

+++ app/controllers/telemechanics_devices_controller.ts
@@ -1,8 +1,10 @@
-import { TUrlParamId } from '#domains/params/types/index';
-import { TelemechanicsDeviceService } from '#services/telemechanics_device_service';
-import { telemechanicsDeviceValidator } from '#validators/telemechanics_device';
-import { urlParamIdValidator } from '#validators/url_params_id';
-import type { HttpContext } from '@adonisjs/core/http';
+import { TUrlParamId } from '#domains/params/types/index'
+import { accessErrorMessages } from '#helpers/access_error_messages'
+import TelemechanicsDevicePolicy from '#policies/telemechanics_device_policy'
+import { TelemechanicsDeviceService } from '#services/telemechanics_device_service'
+import { telemechanicsDeviceValidator } from '#validators/telemechanics_device'
+import { urlParamIdValidator } from '#validators/url_params_id'
+import type { HttpContext } from '@adonisjs/core/http'
 
 export default class TelemechanicsDevicesController {
   async getTelemechanicsDeviceById({ response, params }: HttpContext) {
@@ -12,14 +14,23 @@ export default class TelemechanicsDevicesController {
     return response.status(200).json(telemechanicsDevice)
   }
 
-  async store({ request, response }: HttpContext) {
+  async store({ request, response, bouncer, auth }: HttpContext) {
+    if (await bouncer.with(TelemechanicsDevicePolicy).denies('create')) {
+      return response.status(403).json({ message: accessErrorMessages.create })
+    }
+
+    const { user } = auth
     const validatedData = await request.validateUsing(telemechanicsDeviceValidator)
-    const telemechanicsDevice = await TelemechanicsDeviceService.create(validatedData)
+    const telemechanicsDevice = await TelemechanicsDeviceService.create({ userId: user?.id!, ...validatedData })
 
     return response.status(201).json(telemechanicsDevice)
   }
 
-  async update({ request, response, params }: HttpContext) {
+  async update({ request, response, params, bouncer }: HttpContext) {
+    if (await bouncer.with(TelemechanicsDevicePolicy).denies('edit')) {
+      return response.status(403).json({ message: accessErrorMessages.edit })
+    }
+
     const { id }: TUrlParamId = await urlParamIdValidator.validate(params)
     const validatedData = await request.validateUsing(telemechanicsDeviceValidator)
     const updTelemechanicsDevice = await TelemechanicsDeviceService.update(id, validatedData)
@@ -27,7 +38,11 @@ export default class TelemechanicsDevicesController {
     return response.status(200).json(updTelemechanicsDevice)
   }
 
-  async destroy({ response, params }: HttpContext) {
+  async destroy({ response, params, bouncer }: HttpContext) {
+    if (await bouncer.with(TelemechanicsDevicePolicy).denies('delete')) {
+      return response.status(403).json({ message: accessErrorMessages.delete })
+    }
+
     const { id }: TUrlParamId = await urlParamIdValidator.validate(params)
 
     await TelemechanicsDeviceService.delete(id)
